refactor focus button
